# Conflicts:
#	Client/Bin/Client.exe
#	Client/Bin/Client.pdb
#	Client/Bin/Engine.dll
#	Client/Bin/Engine.pdb
#	Client/Project/Client.vcxproj
#	Client/Project/Client.vcxproj.filters
#	Client/Project/imgui.ini
#	Client/Project/x64/Debug/Client.pch
#	Client/Project/x64/Debug/Client.tlog/CL.command.1.tlog
#	Client/Project/x64/Debug/Client.tlog/CL.read.1.tlog
#	Client/Project/x64/Debug/Client.tlog/CL.write.1.tlog
#	Client/Project/x64/Debug/Client.tlog/Client.write.1u.tlog
#	Client/Project/x64/Debug/Client.tlog/link.command.1.tlog
#	Client/Project/x64/Debug/Client.tlog/link.read.1.tlog
#	Client/Project/x64/Debug/Client.tlog/link.write.1.tlog
#	Client/Project/x64/Debug/vc142.idb
#	Client/Project/x64/Debug/vc142.pdb
#	Client/Source/TestObject.cpp
#	Client/Source/TestScene.cpp
#	Engine/Bin/Engine.dll
#	Engine/Bin/Engine.exp
#	Engine/Bin/Engine.lib
#	Engine/Project/x64/Debug/Engine.tlog/CL.command.1.tlog
#	Engine/Project/x64/Debug/Engine.tlog/CL.read.1.tlog
#	Engine/Project/x64/Debug/Engine.tlog/Engine.write.1u.tlog
#	Engine/Project/x64/Debug/Engine.tlog/link.command.1.tlog
#	Engine/Project/x64/Debug/Engine.tlog/link.read.1.tlog
#	Engine/Project/x64/Debug/Engine.tlog/link.write.1.tlog
#	Engine/Project/x64/Debug/Resources.obj
#	Engine/Project/x64/Debug/vc142.idb
#	Engine/Project/x64/Debug/vc142.pdb
#	Engine/Project/x64/Release/Engine.tlog/CL.14476.write.1.tlog

+++ Engine/Source/RenderTarget.cpp
@@ -40,7 +40,7 @@ void RenderTarget::Initialize(const Info& InitInfo)&
 		}
 	}
 }
-IDirect3DSurface9* RenderTarget::GetSurface(const size_t Index)const&
+IDirect3DSurface9* RenderTarget::GetSurface(const uint64 Index)const&
 {
 	return Surfaces[Index];
 }
